LayoutTests/platform/ios-simulator/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/persisted-user-state-restoration/scroll-restoration-fragment-scrolling-cross-origin-expected.txt

CONSOLE MESSAGE: line 2422: Error: assert_equals: should scroll to fragment expected 800 but got 0


Harness Error (FAIL), message = Error: assert_equals: should scroll to fragment expected 800 but got 0

TIMEOUT Manual scroll restoration should take precedent over scrolling to fragment in cross origin navigation Test timed out
